wParam is already a UINT_PTR, an integer type, so no need to cast it.
Patch by: me

(cherry picked from commit 1e5d8e72b5147bf84598caafc702eb279f3a8b41)

+++ main/vcl/win/source/window/salframe.cxx
@@ -6249,12 +6249,12 @@ LRESULT CALLBACK SalFrameWndProc( HWND hWnd, UINT nMsg, WPARAM wParam, LPARAM lP
             break;
 #if WINVER >= 0x0500
         case WM_IME_REQUEST:
-            if ( static_cast<int>( wParam ) == IMR_RECONVERTSTRING )
+            if ( wParam == IMR_RECONVERTSTRING )
             {
                 nRet = ImplHandleIMEReconvertString( hWnd, lParam );
                 rDef = FALSE;
             }
-	    else if( static_cast<int>( wParam ) == IMR_CONFIRMRECONVERTSTRING )
+	    else if( wParam == IMR_CONFIRMRECONVERTSTRING )
 	    {
 		nRet = ImplHandleIMEConfirmReconvertString( hWnd, lParam );
 		rDef = FALSE;
